#40 implement security

+++ react-client-type/src/components/UserList.tsx
@@ -1,14 +1,20 @@
-import { useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
+import { useEffect, useState, useContext } from 'react';
+import { Link, useNavigate } from 'react-router-dom';
 
 import Table from 'react-bootstrap/Table';
 import Button from 'react-bootstrap/Button';
 import Container from 'react-bootstrap/Container';
 import AdminMenuBar from './AdminMenuBar';
 
+import AuthContext from '../AuthContext';
+
 
 function UserList() {
   const [userAccounts, setUserAccounts] = useState([]);
+
+  const auth = useContext(AuthContext);
+
+  const navigate = useNavigate();
 
   useEffect(() => {
     fetch('http://localhost:8080/api/useraccount')
@@ -23,37 +29,42 @@ function UserList() {
       .catch(console.log);
   }, []);
 
-  // TODO handleDeleteUser
   const handleDeleteUser = (appUserId: number) => {
+    if (auth === undefined || auth.user === null) {
+      window.alert('You must be logged in to access this feature')
+      navigate('/login');
+      return;
+    }
+
     const userAccount: any = userAccounts.find(userAccount => userAccount['appUserId'] === appUserId);
 
-    if(window.confirm(
-    `    Deletion is permanent.
+    if (window.confirm(
+      `    Deletion is permanent.
     Are you sure you want to proceded?
     Delete user ${userAccount['firstName']} ${userAccount['lastName']}?`)) {
       const init = {
         method: 'DELETE',
-        // headers: {
-        //   'Authorization': `Bearer ${auth.user.token}`
-        // },
+        headers: {
+          'Authorization': `Bearer ${auth.user.token}`
+        },
       };
 
       fetch(`http://localhost:8080/api/useraccount/${appUserId}`, init)
-      .then(response => {
-        if (response.status === 204) {
-          const newUserAccount = userAccounts.filter(userAccount => userAccount['appUserId'] !== appUserId);
-          setUserAccounts(newUserAccount);
-        } else {
-          return Promise.reject(`Unexpected status code: ${response.status}`);
-        }
-      })
-      .catch(console.log);
+        .then(response => {
+          if (response.status === 204) {
+            const newUserAccount = userAccounts.filter(userAccount => userAccount['appUserId'] !== appUserId);
+            setUserAccounts(newUserAccount);
+          } else {
+            return Promise.reject(`Unexpected status code: ${response.status}`);
+          }
+        })
+        .catch(console.log);
     }
   };
 
   return (
     <>
-    <AdminMenuBar />
+      <AdminMenuBar />
       <h1 className="text-center">All Customers We Work With</h1>
       <Container>
         <Table>
@@ -69,27 +80,27 @@ function UserList() {
             </tr>
           </thead>
           <tbody>
-          {userAccounts.map(userAccount => (
-            <tr key={userAccount['appUserId']}>
-              <td>{userAccount['appUserId']}</td>
-              <td>{userAccount['firstName']} {userAccount['lastName']}</td>
-              <td>{userAccount['email']}</td>
-              <td>{userAccount['address']}</td>
-              <td>{userAccount['phone']}</td>
-              <td>{userAccount['dob']}</td>
-              <td>
-                <div className="float-right mr-2">
-                  <Link className="btn btn-primary btn-sm mr-2 me-2" to={`/userupdateform/${userAccount['appUserId']}`}>
-                    <i className="bi bi-pencil-square"></i>
-                  </Link>
+            {userAccounts.map(userAccount => (
+              <tr key={userAccount['appUserId']}>
+                <td>{userAccount['appUserId']}</td>
+                <td>{userAccount['firstName']} {userAccount['lastName']}</td>
+                <td>{userAccount['email']}</td>
+                <td>{userAccount['address']}</td>
+                <td>{userAccount['phone']}</td>
+                <td>{userAccount['dob']}</td>
+                <td>
+                  <div className="float-right mr-2">
+                    <Link className="btn btn-primary btn-sm mr-2 me-2" to={`/userupdateform/${userAccount['appUserId']}`}>
+                      <i className="bi bi-pencil-square"></i>
+                    </Link>
                     <Button variant="danger" className="btn-sm" onClick={() => handleDeleteUser(userAccount['appUserId'])}>
                       <i className="bi bi-trash"></i>
                     </Button>
-                </div>
-              </td>
-            </tr>
-          ))}
-        </tbody>
+                  </div>
+                </td>
+              </tr>
+            ))}
+          </tbody>
         </Table>
         <Button>Edit Customer</Button>
       </Container>
